fix(tools/archtest): err no-cache + testing.Short() + move helper — PR-A49 reviewer P1 Cx2/P2 Cx1
P1 Cx2 (1): CellsTopicResolver — replace sync.Once (caches errors permanently)
with sync.Mutex pattern: success is cached, error is NOT cached so next call
can retry. Avoids permanently poisoning the singleton on transient toolchain
failures.

P1 Cx2 (2): TestSecurityTopicsDoNotOptInFailOpen — add testing.Short() skip
guard. packages.Load("./cells/...") costs ~5-10s; -short lets developers skip
module-wide archtest locally. CI does not pass -short; behaviour unchanged.

P2 Cx1 (3): findCellProductionGoFiles — moved from outbox_topic_test.go to
helpers_test.go (new file). The function is called from both outbox_topic_test.go
and repoerr_test.go; living in a shared helpers file matches the "shared test
helper" convention and removes the cross-file dependency hidden in one file.
